Fetch all tables and databases across pages via pageToken in LanceRestCatalog
